Playground development

+++ Sources/SwiftProcessing/Core/Image.swift
@@ -257,7 +257,7 @@ open class Image {
             }else if filterType == Sketch.MONOCHROME{
                 filter = CIFilter(name: "CIColorMonochrome")
                 let c = params as! Color
-                let ciColor = CIColor(red: c.red, green: c.green, blue: c.blue)
+                let ciColor = CIColor(red: CGFloat(c.red), green: CGFloat(c.green), blue: CGFloat(c.blue))
                 filter?.setValue(currentCIImage, forKey: kCIInputImageKey)
                 filter?.setValue(ciColor, forKey: kCIInputColorKey)
             }else if filterType == Sketch.INVERT{
